🦄 refactor(Function Split):
Functions are now in their own files to be called when necessary
